requirements.txt
streamlit==1.38.0
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain_chroma==0.1.3
langchain_cohere==0.2.4
langchain_community==0.2.16
langchain_google_genai==1.0.10
langchain_core==0.2.39